requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
email-validator==2.3.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.0

# Logging
structlog==23.2.0

# CLI
click==8.1.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
requests==2.31.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Optional: Redis for advanced rate limiting
# redis==5.0.1
# aioredis==2.0.1
